Added puzzle
created a puzzle class to help the organization of the code.
