+++ braukhoffna_rennerj_ werneral_prog1/.idea/misc.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="JavaScriptSettings">
+    <option name="languageLevel" value="ES6" />
+  </component>
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.7 (braukhoffna_rennerj_ werneral_prog1)" project-jdk-type="Python SDK" />
+</project>
